cursos/views.py

from rest_framework.views import APIView
from rest_framework.response import Response

from .models import Curso, Avaliacao
from .serializers import CursoSerializer, AvaliacaoSerializer

class CursoAPIView(APIView):

    def get(self, request):

        cursos = Curso.objects.all()
        serializer = CursoSerializer(cursos, many=True)

        return Response(serializer.data)

class AvaliacaoAPIView(APIView):

    def get(self, request):
        
        avaliacaoes = Avaliacao.objects.all()
        serializer = AvaliacaoSerializer(avaliacaoes, many=True)

        return Response(serializer.data)