requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==3.1.3
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
plotly==4.12.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2020.4
retrying==1.3.3
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sqlparse==0.4.1
